blog_backend/requirements.txt
asgiref==3.3.0
Django==3.1.3
django-cors-headers==3.5.0
django-summernote==0.8.11.6
djangorestframework==3.12.1
Pillow==8.0.1
pkg-resources==0.0.0
pytz==2020.4
sqlparse==0.4.1
